jobs/ingest_kolkostruva/pyproject.toml
[project]
name = "spesto-ingest-kolkostruva"
version = "0.1.0"
description = "Daily fetch of kolkostruva.bg open-data price feeds into the GCS landed zone."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "google-cloud-storage>=2.18",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
